+++ src/Lib/Routing.php
@@ -6,9 +6,7 @@ use Foodsharing\Modules\Basket\BasketXhr;
 use Foodsharing\Modules\BusinessCard\BusinessCardControl;
 use Foodsharing\Modules\Dashboard\DashboardControl;
 use Foodsharing\Modules\Index\IndexControl;
-use Foodsharing\Modules\Logout\LogoutControl;
 use Foodsharing\Modules\Mailbox\MailboxControl;
-use Foodsharing\Modules\Relogin\ReloginControl;
 use Foodsharing\Modules\Store\StoreController;
 use Foodsharing\Modules\StoreUser\StoreUserControl;
 use Foodsharing\Modules\Voting\VotingControl;
@@ -31,11 +29,9 @@ class Routing
         'bcard' => 'BusinessCard',
         'dashboard' => 'Dashboard',
         'index' => 'Index',
-        'logout' => 'Logout',
         'mailbox' => 'Mailbox',
         'poll' => 'Voting',
         'register' => 'Register',
-        'relogin' => 'Relogin',
         'report' => 'Report',
         'search' => 'Search',
         'betrieb' => 'Store',
@@ -49,10 +45,8 @@ class Routing
         'bcard' => BusinessCardControl::class,
         'dashboard' => DashboardControl::class,
         'index' => IndexControl::class,
-        'logout' => LogoutControl::class,
         'mailbox' => MailboxControl::class,
         'poll' => VotingControl::class,
-        'relogin' => ReloginControl::class,
         'fsbetrieb' => StoreUserControl::class,
         'groups' => WorkGroupControl::class,
         'store' => StoreController::class,
@@ -82,6 +76,8 @@ class Routing
         'register',
         'application',
         'msg',
+        'logout',
+        'relogin',
     ];
 
     private const RENAMES = [