+++ CHANGELOG.md
@@ -82,6 +82,7 @@
 - Replace TinyMCE with Quill #2141 !3660 @SaiCode
 - Refactored add and edit form to vue and removed jquery tagedit #2137 !3520 @chriswalg @alex.simm
 - Remove all usages of deprecated columns email_pass and email from fs_bezirk #1853 !3710 @AntonBallmaier
+- Port the logout and relogin controllers to Symfony !3682 @alex.simm
 
 ## Dev/Test/CI stuff
 - Improved API tests runtime !3506 @alex.simm